Calendar Model and TeamCalendarEntityManager
Changed CalendarStructure to CalendarModel and made a
TeamCalendarEntityManager. Wrote toJSON and fromJSON functions for
TeamCalendar, and stubbed them out for PersonalCalendar. Made changes to
the ManagerLayer to add the TeamCalendarEntityManager as the tutorial in
the wiki described, but I am still getting errors there.
TeamCalendarEntityManager will still need an update function at the
least, and the functions that we don't want to be used will need to
throw exceptions.